Fix the main layout issue with the list of items
The main issue appears to come from having  `HStack`s with `Toggle`
(all inside a `List`, not sure if that matters though).

This commit sidesteps the issue by using checkboxes instead of toggles.

There are still issues (stuff disappears and/or has the wrong position
when going out of and into view), but at least they aren't completely
breaking anymore.